requirements.txt
streamlit
pillow
google-generativeai
plantdoc-predictor
